# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Spec for footbal team management",
        "title": "API Football Team Management",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "afrizalsebastian",
            "email": "sebastiangurning@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/api/v1/teams": {
            "get": {
                "description": "Get List Team",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Teams"
                ],
                "summary": "Get List team",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-array_dto_GetListTeamItem"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-any"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new team",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Teams"
                ],
                "summary": "Create Team",
                "parameters": [
                    {
                        "description": "Create team body request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateTeamRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-dto_CreateTeamResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-any"
                        }
                    }
                }
            }
        },
        "/api/v1/teams/{teamId}": {
            "get": {
                "description": "Get Detail Team",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Teams"
                ],
                "summary": "Get Detail Team",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team ID",
                        "name": "teamId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-dto_GetTeamDetail"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-any"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update Team Data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Teams"
                ],
                "summary": "Update Team Data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team ID",
                        "name": "teamId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update team request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateTeamRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-dto_GetTeamDetail"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-any"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete team (soft-delete)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Teams"
                ],
                "summary": "Delete team (soft-delete)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team ID",
                        "name": "teamId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-any"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-any"
                        }
                    }
                }
            }
        },
        "/api/v1/teams/{teamId}/players": {
            "get": {
                "description": "Get List Team Player",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Teams"
                ],
                "summary": "Get List Team Player",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team ID",
                        "name": "teamId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-array_dto_GetListTeamPlayerItem"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-any"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new team Player",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Teams"
                ],
                "summary": "Create Team Player",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team ID",
                        "name": "teamId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Create team bory request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreatePlayerTeamRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-dto_CreatePlayerTeamResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.WebResponse-any"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.ErrorsDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "api.WebResponse-any": {
            "type": "object",
            "properties": {
                "data": {},
                "error_message": {
                    "type": "string"
                },
                "errors_detail": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.ErrorsDetail"
                    }
                },
                "http_code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "message_code": {
                    "type": "integer"
                },
                "message_id": {
                    "type": "string"
                }
            }
        },
...
        "api.WebResponse-array_dto_GetListTeamItem": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.GetListTeamItem"
                    }
                },
                "error_message": {
                    "type": "string"
                },
                "errors_detail": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.ErrorsDetail"
                    }
                },
                "http_code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "message_code": {
                    "type": "integer"
                },
                "message_id": {
                    "type": "string"
                }
            }
        },
        "api.WebResponse-array_dto_GetListTeamPlayerItem": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.GetListTeamPlayerItem"
                    }
                },
                "error_message": {
                    "type": "string"
                },
                "errors_detail": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.ErrorsDetail"
                    }
                },
                "http_code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "message_code": {
                    "type": "integer"
                },
                "message_id": {
                    "type": "string"
                }
            }
        },
...
        "api.WebResponse-dto_CreatePlayerTeamResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.CreatePlayerTeamResponse"
                },
                "error_message": {
                    "type": "string"
                },
                "errors_detail": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.ErrorsDetail"
                    }
                },
                "http_code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "message_code": {
                    "type": "integer"
                },
                "message_id": {
                    "type": "string"
                }
            }
        },
        "api.WebResponse-dto_CreateTeamResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.CreateTeamResponse"
                },
                "error_message": {
                    "type": "string"
                },
                "errors_detail": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.ErrorsDetail"
                    }
                },
                "http_code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "message_code": {
                    "type": "integer"
                },
                "message_id": {
                    "type": "string"
                }
            }
        },
...
        "api.WebResponse-dto_GetTeamDetail": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/dto.GetTeamDetail"
                },
                "error_message": {
                    "type": "string"
                },
                "errors_detail": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.ErrorsDetail"
                    }
                },
                "http_code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "message_code": {
                    "type": "integer"
                },
                "message_id": {
                    "type": "string"
                }
            }
        },
...
        "dto.CreatePlayerTeamRequest": {
            "type": "object",
            "required": [
                "Position",
                "height",
                "jersey_number",
                "name",
                "weight"
            ],
            "properties": {
                "Position": {
                    "type": "string"
                },
                "height": {
                    "type": "number"
                },
                "jersey_number": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "weight": {
                    "type": "number"
                }
            }
        },
        "dto.CreatePlayerTeamResponse": {
            "type": "object",
            "properties": {
                "height": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "jersey_number": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "position": {
                    "type": "string"
                },
                "team_id": {
                    "type": "string"
                },
                "weight": {
                    "type": "number"
                }
            }
        },
        "dto.CreateTeamRequest": {
            "type": "object",
            "required": [
                "address",
                "city",
                "founded_year",
                "logo",
                "name"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "founded_year": {
                    "type": "string"
                },
                "logo": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "dto.CreateTeamResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "founded_year": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "logo": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "dto.GetListTeamItem": {
            "type": "object",
            "properties": {
                "founded_year": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "logo": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "dto.GetListTeamPlayerItem": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "jersey_number": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "position": {
                    "$ref": "#/definitions/dto.PlayerPosition"
                },
                "team_id": {
                    "type": "string"
                }
            }
        },
...
        "dto.GetTeamDetail": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "founded_year": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_deleted": {
                    "type": "boolean"
                },
                "logo": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "dto.PlayerPosition": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "dto.UpdateTeamRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "founded_year": {
                    "type": "string"
                },
                "logo": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Please input token with format: Bearer __TOKEN__",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}